+++ util_disk_linux_test.go
@@ -1,0 +1,11 @@
+package psn
+
+import (
+	"fmt"
+	"testing"
+)
+
+func TestGetDevice(t *testing.T) {
+	s, err := GetDevice("/")
+	fmt.Println(s, err)
+}
