+++ util_net_linux_test.go
@@ -1,0 +1,11 @@
+package psn
+
+import (
+	"fmt"
+	"testing"
+)
+
+func TestGetDefaultInterface(t *testing.T) {
+	s, err := GetDefaultInterface()
+	fmt.Println(s, err)
+}
